+++ src/main/java/garlicbears/quiz/domain/game/admin/controller/AdminTopicController.java
@@ -1,7 +1,5 @@
 package garlicbears.quiz.domain.game.admin.controller;
 
-import static org.springframework.util.ClassUtils.*;
-
 import java.util.Map;
 import java.util.Optional;
