+++ src/main/java/garlicbears/quiz/domain/game/admin/service/TopicService.java
@@ -17,7 +17,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import garlicbears.quiz.domain.common.entity.Active;
 import garlicbears.quiz.domain.common.entity.Image;
-import garlicbears.quiz.domain.common.entity.User;
 import garlicbears.quiz.domain.game.admin.dto.ResponseQuestionDto;
 import garlicbears.quiz.domain.game.admin.dto.ResponseTopicDto;
 import garlicbears.quiz.domain.game.admin.dto.ResponseTopicListDto;
