+++ src/main/java/garlicbears/quiz/domain/management/common/repository/ImageRepository.java
@@ -1,7 +1,5 @@
 package garlicbears.quiz.domain.management.common.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import garlicbears.quiz.domain.common.entity.Image;
